==> Bloc1&2/Definitive files/8-CleanCSVWithTags.ipynb <==
**Check if all columns have all rows with the same type**

import pandas as pd

# Replace 'your_file.csv' with the actual file path
df = pd.read_csv('TEST\data_clean3_with_SteamTags.csv')

unique_data_types = {}

for column in df.columns:
    data_types = set(df[column].apply(type))
    unique_data_types[column] = data_types

for column, data_types in unique_data_types.items():
    print(f"Column: {column}, Unique Data Type(s): {data_types}")
